добавил ммногопоточность и класс базы данных

+++ src/main/kotlin/MenuItem.kt
@@ -1,5 +1,5 @@
 data class MenuItem(
     val name: String,
     var price: Double,
-    var complexity: Int
+    var complexity: Long // заполняются минуты
 )
